Extraction of data classes and enums with namespace modification
