+++ .gitignore
@@ -24,6 +24,7 @@ ExportedObj/
 *.pdb
 *.opendb
 *.VC.db
+.idea
 
 # Unity3D generated meta files
 *.pidb.meta
