+++ Runtime/KredPurchaseType.cs
@@ -1,0 +1,8 @@
+namespace Kongregate.Web {
+    public enum KredPurchaseType
+    {
+        Default,
+        Offers,
+        Mobile,
+    }
+}
